Criação do JWT e rotas autenticadas feita no backend

+++ server/src/routes/memories.ts
@@ -1,36 +1,47 @@
 import { FastifyInstance } from 'fastify'
-import {z} from "zod"
+import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook("preHandler", async (req) => {
+    await req.jwtVerify()
+  })
+
+  app.get('/memories', async (req) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: req.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
     })
 
-    return memories.map(memory => {
-      return{
+    return memories.map((memory) => {
+      return {
         id: memory.id,
         coverUrl: memory.coverUrl,
-        excerpt: memory.Content.substring(0, 115).concat("...")
+        excerpt: memory.Content.substring(0, 115).concat('...'),
       }
     })
   })
 
-  app.get('/memories/:id', async (req) => {
+  app.get('/memories/:id', async (req, reply) => {
     const paramsSchema = z.object({
-      id: z.string().uuid()
+      id: z.string().uuid(),
     })
 
-    const {id} = paramsSchema.parse(req.params)
+    const { id } = paramsSchema.parse(req.params)
 
     const memory = await prisma.memory.findUniqueOrThrow({
-      where:{
+      where: {
         id,
-      }
+      },
     })
+
+    if (!memory.isPublic && memory.userId !== req.user.sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -39,39 +50,49 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const bodySchema = z.object({
       Content: z.string(),
       coverUrl: z.string(),
-      isPublic: z.coerce.boolean().default(false)
+      isPublic: z.coerce.boolean().default(false),
     })
 
-    const { Content,coverUrl, isPublic} = bodySchema.parse(req.body)
+    const { Content, coverUrl, isPublic } = bodySchema.parse(req.body)
 
     const memory = await prisma.memory.create({
-      data:{
+      data: {
         Content,
         coverUrl,
         isPublic,
-        userId:"882478db-abdc-470c-90d8-97daeac959ca"
-      }
+        userId: req.user.sub
+      },
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (req) => {
+  app.put('/memories/:id', async (req, reply) => {
     const paramsSchema = z.object({
-      id: z.string().uuid()
+      id: z.string().uuid(),
     })
 
-    const {id} = paramsSchema.parse(req.params)
+    const { id } = paramsSchema.parse(req.params)
 
     const bodySchema = z.object({
       Content: z.string(),
       coverUrl: z.string(),
-      isPublic: z.coerce.boolean().default(false)
+      isPublic: z.coerce.boolean().default(false),
     })
 
-    const { Content,coverUrl, isPublic} = bodySchema.parse(req.body)
+    const { Content, coverUrl, isPublic } = bodySchema.parse(req.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id
+      }
+    })
+
+    if (memory.userId !== req.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -79,24 +100,33 @@ export async function memoriesRoutes(app: FastifyInstance) {
         Content,
         coverUrl,
         isPublic,
-      }
+      },
     })
 
     return memory
-
   })
 
-  app.delete('/memories/:id', async (req) => {
+  app.delete('/memories/:id', async (req, reply) => {
     const paramsSchema = z.object({
-      id: z.string().uuid()
+      id: z.string().uuid(),
     })
 
-    const {id} = paramsSchema.parse(req.params)
+    const { id } = paramsSchema.parse(req.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id
+      }
+    })
+
+    if (memory.userId !== req.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
-      where:{
+      where: {
         id,
-      }
+      },
     })
   })
 }
